changed amount to uint64 for native tokens

+++ everscale/contracts/bridge/event-contracts/multivault/MultiVaultEverscaleSolanaEventNative.sol
@@ -22,7 +22,7 @@ contract MultiVaultEverscaleSolanaEventNative is EverscaleSolanaBaseEvent, IMult
     address tokenWallet;
     address token;
     address remainingGasTo;
-    uint128 amount;
+    uint64 amount;
     uint256 recipient;
 
     string name;
@@ -65,7 +65,7 @@ contract MultiVaultEverscaleSolanaEventNative is EverscaleSolanaBaseEvent, IMult
             recipient
         ) = abi.decode(
             eventInitData.voteData.eventData,
-            (address, address, address, address, uint128, uint256)
+            (address, address, address, address, uint64, uint256)
         );
 
         ITokenRoot(token).name{
@@ -138,7 +138,7 @@ contract MultiVaultEverscaleSolanaEventNative is EverscaleSolanaBaseEvent, IMult
         address tokenWallet_,
         address token_,
         address remainingGasTo_,
-        uint128 amount_,
+        uint64 amount_,
         uint256 recipient_,
         string name_,
         string symbol_,
