+++ everscale/contracts/bridge/event-contracts/multivault/MultiVaultSolanaEverscaleEventNative.sol
@@ -15,7 +15,7 @@ import "./../base/SolanaEverscaleBaseEvent.sol";
 
 contract MultiVaultSolanaEverscaleEventNative is SolanaEverscaleBaseEvent, IMultiVaultSolanaEverscaleEventNative {
     address token;
-    uint128 amount;
+    uint64 amount;
     address recipient;
 
     address proxy;
@@ -51,7 +51,7 @@ contract MultiVaultSolanaEverscaleEventNative is SolanaEverscaleBaseEvent, IMult
             recipient_addr
         ) = abi.decode(
             eventInitData.voteData.eventData,
-            (int8, uint256, uint128, int8, uint256)
+            (int8, uint256, uint64, int8, uint256)
         );
 
         token = address.makeAddrStd(token_wid, token_addr);
@@ -90,7 +90,7 @@ contract MultiVaultSolanaEverscaleEventNative is SolanaEverscaleBaseEvent, IMult
 
     function getDecodedData() external override responsible returns(
         address token_,
-        uint128 amount_,
+        uint64 amount_,
         address recipient_,
         address proxy_,
         address tokenWallet_
